Add files via upload
A file sharing app with QR code using Python would offer the same basic functionality as a traditional file sharing app, with the added convenience of using QR codes to share files.

Here's how such an app might work:

Uploading and Storing Files:
The user would first upload a file to the app via the web-based interface. The app would then store the file on the server and generate a unique QR code that represents the file.

Sharing the QR Code:
The user could then share the QR code with other users via email, text message, or other messaging platforms. The recipient would scan the QR code with their device's camera, which would take them to a download link for the file.

Downloading Files:
Once the recipient clicks on the download link, the file would be downloaded to their device.

To implement this functionality, the app would use a Python web framework such as Flask or Django, as well as a QR code library like PyQRCode or QRcode. The app would also likely use a database to store information about uploaded files and associated QR codes.